+++ src/resource/image.ts
@@ -10,4 +10,5 @@ export default {
   product_box: require("./media/images/product-box.png"),
   empty_shop: require("./media/images/empty_shop.png"),
   emptyPromotion: require("./media/images/default_promotion.jpg"),
+  emptyProFreebie: require("./media/images/IconEmtryStateOnFreebie.png"), 
 };